Force models to keep all levels and keep metadata
`lm()` automatically drops unused factor levels, causing mismatches in a decomposition's sub-models (especially when coefficients baselines are adjusted).  Pre-expanding the model matrix (in effect, creating dummies) and calling `lm()` on this X matrix forces it to keep all the levels.

The names of estimated coefficients are sensitive to the `lm()` call.  The new calls should keep the old names for everything except the intercept (which can be cleaned up later).

Unfortunately, some of the attributes we use elsewhere aren't retained by these new `lm()` calls, so this commit saves several of the original attributes along with the fitted models.